requirements.txt
ccxt>=4.2.0
pandas>=2.2.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
jupyter>=1.0.0
matplotlib>=3.8.0
streamlit>=1.30.0
streamlit-autorefresh>=1.0.0
plotly>=5.18.0
